Skeleton of upload functionality
Currently mounted at /publish this commit allows users to
upload files and/or urls and see what they have currently uploaded
before answering questions as to where they want to put the files.

All data is currently in session cookies (booo) and doesn't yet
contain the publisher (perhaps if we launch from the publisher page
we can store it). At some point, it may be useful to store this
information in a PublishSession model so that they can be resumed.

The PublishController needs to spawn processes as soon as it is able to
do any processing on the url or file. IFF it has completed by the time
that the user reaches the /publish/find page then we can do a bit more
magic.  Otherwise the /publish/find page should cancel the processes we
spawned earlier.