Seeds the database with some templated datasets for the seeded publishers.
Each dataset currently has a fake ```datafile``` which contains
information about the file, such as the URL, name and hopefully
a description of the dataset.

There are also changes to the publisher index templates to show the
datasets for that publisher, updates on the dataset show page to include
the datafiles etc.