Some changes to the publisher index+show pages
Remove some stuff from the publisher pages and add temporary auth so we
can model the user being able to add data (or not) to a particular
publisher.

Makes assumptions that the publisher workflow can take either a
publisher or a dataset (in which case the workflow is really short).