Add some stats to the publisher page intended for drill-down.
Adds some potentially useful numbers (some are made up) for each
publisher home page.  Each of the numbers in the statistics section
should be a link, but the actual link will need another PR after we have
built the drill-down pages.

This may not be the correct place for this info, but we can always move
it later.