Run all incoming CSV data through iconv
As a temporary fix to #125 now when we download a CSV for preview, we'll
write it to disk, run it through iconv to convert to utf8 (assuming
source is iso 8859 1 - which it normally is, thanks Excel) and then send
that instead.

In future, we don't want to do this every time we fetch the CSV, we
should cache the converted copy.